training shell

+++ elmo.py
@@ -1,15 +1,9 @@
 # Need allennlp with torch==0.3.1 && GPU
-
+import os
+import sys
 import pickle
 import re
 import unicodedata
-
-from allennlp.modules.elmo import Elmo, batch_to_ids
-
-options_file = "https://s3-us-west-2.amazonaws.com/allennlp/models/elmo/2x4096_512_2048cnn_2xhighway/elmo_2x4096_512_2048cnn_2xhighway_options.json" 
-weight_file = "https://s3-us-west-2.amazonaws.com/allennlp/models/elmo/2x4096_512_2048cnn_2xhighway/elmo_2x4096_512_2048cnn_2xhighway_weights.hdf5"
-
-elmo = Elmo(options_file, weight_file, 2, dropout=0)
 
 def unicodeToAscii(s):
     return ''.join(
@@ -32,14 +26,14 @@ def loadConll(path='CoNLL_data/train.txt'):
     pairs = [[[e for e in normalizeString(s).split(' ')] for s in l.split('\t')] for l in lines]
     return pairs
 
-def elmoFromPair(pair):
+def elmoFromPair(pair, elmo):
     # choose layer 2 representations
     layer = 1
     character_ids = batch_to_ids(pair)
     input_tensor, output_tensor = elmo(character_ids)['elmo_representations'][layer].data
     return (input_tensor, output_tensor)
 
-def sen2elmo(sentence):
+def sen2elmo(sentence, elmo):
     layer = 1
     character_ids = batch_to_ids(sentence)
     tensor = elmo(character_ids)['elmo_representations'][layer][0].data
@@ -47,23 +41,40 @@ def sen2elmo(sentence):
 
 # input_lang, output_lang, pairs = prepareData('eng', 'fra', True)
 if __name__ == '__main__':
-    # TEST SMALL DATASET
-    # path = 'CoNLL_data/train.txt'
-    # emb_path = 'CoNLL_data/train_small.elmo'
-    # pairs = loadConll(path)[:10]
+    # python elmo.py train.txt train.elmo 
+    if len(sys.argv) == 2:
+        sentence_path = sys.argv[1]
+        emb_path = sys.argv[2]
+    else:
+        # TEST SMALL DATASET
+        sentence_path = 'CoNLL_data/train.txt'
+        emb_path = 'CoNLL_data/train_small.elmo'
+        pairs = loadConll(sentence_path)[:10]
+
+    # no need to pretrain again
+    if os.path.isfile(emb_path): 
+        exit()
+
+    from allennlp.modules.elmo import Elmo, batch_to_ids
+    options_file = "https://s3-us-west-2.amazonaws.com/allennlp/models/elmo/2x4096_512_2048cnn_2xhighway/elmo_2x4096_512_2048cnn_2xhighway_options.json" 
+    weight_file = "https://s3-us-west-2.amazonaws.com/allennlp/models/elmo/2x4096_512_2048cnn_2xhighway/elmo_2x4096_512_2048cnn_2xhighway_weights.hdf5"
+    elmo = Elmo(options_file, weight_file, 2, dropout=0)
 
     # BASELINE
-    # path = 'CoNLL_data/baseline_train.txt'
+    # sentence_path = 'CoNLL_data/baseline_train.txt'
     # emb_path = 'CoNLL_data/train_baseline.elmo'
 
     # WITH ERROR TAG
-    path = 'CoNLL_data/train.txt'
-    emb_path = 'CoNLL_data/train.elmo'
+    # sentence_path = 'CoNLL_data/train.txt'
+    # emb_path = 'CoNLL_data/train.elmo'
 
     # pairs = [[['First', 'sentence', '.'], ['Another', '.']]]
-    pairs = loadConll(path)
+    pairs = loadConll(sentence_path)
 
     # (elmo, text) pairs
-    embeddings = [[sen2elmo(pair[0]), pair[1]] for pair in pairs]
+    # embeddings = [[sen2elmo(pair[0], elmo), pair[1]] for pair in pairs]
+    # (elmo, elmo) pairs
+    embeddings = [elmoFromPair(pair, elmo) for pair in pairs]
+
     with open(emb_path, 'wb') as file:
         pickle.dump(embeddings, file)
